requirements.txt
appscript==1.2.2
black==23.3.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docx2pdf==0.1.8
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
lxml==4.9.2
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
tomli==2.0.1
tqdm==4.65.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
